// frontend/src/api/model/registryCastEntry.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */

export interface RegistryCastEntry {
    name?: string;
    aliases?: string[];
    routing?: string;
    confirmed?: boolean;
    matchState?: string;
    /** @nullable */
    proposedPersonaId?: string | null;
    /** @nullable */
    proposedPersonaName?: string | null;
    /** @nullable */
    matchedPersonaId?: string | null;
}